Resolving compilation errors
